Added 'all' option to --platform (#206)
* patched to use --platform flag

* Fixed bug where vagrant tried to find a -all host

* Fixed more bugs related to using multiple platforms

* fixed state file from incorrectly updating

* YAPF'ed

+++ molecule/provisioners.py
@@ -22,7 +22,7 @@ import abc
 import collections
 import json
 import os
-
+import copy
 import colorama
 import docker
 import vagrant
@@ -218,6 +218,7 @@ class VagrantProvisioner(BaseProvisioner):
         molecule._env['VAGRANT_VAGRANTFILE'] = molecule._config.config[
             'molecule']['vagrantfile_file']
         self._vagrant.env = molecule._env
+        self._updated_multiplatform = False
 
     def _get_provider(self):
         if self.m._args.get('--provider'):
@@ -235,10 +236,12 @@ class VagrantProvisioner(BaseProvisioner):
 
     def _get_platform(self):
         if self.m._args.get('--platform'):
-            if not [item
-                    for item in self.m._config.config['vagrant']['platforms']
-                    if item['name'] == self.m._args['--platform']]:
-                raise InvalidPlatformSpecified()
+            if self.m._args['--platform'] != 'all':
+                if not [item
+                        for item in self.m._config.config['vagrant'][
+                            'platforms']
+                        if item['name'] == self.m._args['--platform']]:
+                    raise InvalidPlatformSpecified()
             self.m._state['default_platform'] = self.m._args['--platform']
             self.m._env['MOLECULE_PLATFORM'] = self.m._args['--platform']
         else:
@@ -261,6 +264,7 @@ class VagrantProvisioner(BaseProvisioner):
 
     @property
     def instances(self):
+        self.populate_platform_instances()
         return self.m._config.config['vagrant']['instances']
 
     @property
@@ -354,15 +358,18 @@ class VagrantProvisioner(BaseProvisioner):
         return params
 
     def up(self, no_provision=True):
+        self.populate_platform_instances()
         self._write_vagrant_file()
         self._vagrant.up(no_provision)
 
     def destroy(self):
-        self._write_vagrant_file()
         if self.m._state.get('created'):
             self._vagrant.destroy()
 
-        os.remove(self.m._config.config['molecule']['vagrantfile_file'])
+        if os._exists(self.m._config.config['molecule']['vagrantfile_file']):
+            os.remove(self.m._config.config['molecule']['vagrantfile_file'])
+
+        self.m._state['platformcloned'] = False
 
     def status(self):
         return self._vagrant.status()
@@ -377,8 +384,13 @@ class VagrantProvisioner(BaseProvisioner):
         # TODO: for Ansiblev2, the following line must have s/ssh_//
         template = '{} ansible_ssh_host={} ansible_ssh_port={} ansible_ssh_private_key_file={} ansible_ssh_user={}\n'
 
-        ssh = self.conf(vm_name=utilities.format_instance_name(
-            instance['name'], self._platform, self.instances))
+        if not self._updated_multiplatform:
+            ssh = self.conf(vm_name=utilities.format_instance_name(
+                instance['name'], self._platform, self.instances))
+        else:
+            ssh = self.conf(vm_name=utilities.format_instance_name(
+                instance['name'], 'all', self.instances))
+
         return template.format(ssh['Host'], ssh['HostName'], ssh['Port'],
                                ssh['IdentityFile'], ssh['User'])
 
@@ -394,6 +406,27 @@ class VagrantProvisioner(BaseProvisioner):
             conf['HostName'], conf['User'], conf['Port'], conf['IdentityFile'],
             ' '.join(self.m._config.config['molecule']['raw_ssh_args'])
         ]
+
+    def populate_platform_instances(self):
+        if self.m._args.get('--platform'):
+            if len(self.m._config.config['vagrant'][
+                    'platforms']) > 1 and self.m._args[
+                        '--platform'] == 'all' and not self._updated_multiplatform:
+                new_instances = []
+
+                for instance in self.m._config.config['vagrant']['instances']:
+                    for platform in self.m._config.config['vagrant'][
+                            'platforms']:
+                        platform_instance = copy.deepcopy(instance)
+                        platform_instance['platform'] = platform['box']
+                        platform_instance['name'] = instance[
+                            'name'] + '-' + platform['name']
+                        platform_instance['vm_name'] = instance[
+                            'name'] + '-' + platform['name']
+                        new_instances.append(platform_instance)
+
+                self.m._config.config['vagrant']['instances'] = new_instances
+                self._updated_multiplatform = True
 
 
 # Place holder for Proxmox, partially implemented
